+++ src/client/cypress/e2e/detailPage/labeling.cy.js
@@ -2,6 +2,7 @@ import { evaluateCoordinate, evaluateSelect, hasAiStyle, hasError, isDisabled } 
 import {
   newEditableBorehole,
   newUneditableBorehole,
+  selectInputFile,
   startBoreholeEditing,
   stopBoreholeEditing,
 } from "../helpers/testHelpers.js";
@@ -113,15 +114,7 @@ describe("Test labeling tool", () => {
     isFileActive("borehole_attachment_1.pdf", false);
     isFileActive("borehole_attachment_3.pdf", true);
 
-    const crypto = window.crypto || window.msCrypto;
-    cy.get("input[type=file]").selectFile(
-      {
-        contents: Cypress.Buffer.from(crypto.getRandomValues(new Uint32Array(1)).toString()),
-        fileName: "WOLFHEART.pdf",
-        mimeType: "application/pdf",
-      },
-      { force: true },
-    );
+    selectInputFile("WOLFHEART.pdf", "application/pdf");
 
     cy.get('[data-cy="labeling-file-button-select"]').contains("WOLFHEART.pdf");
     cy.get('[data-cy="button-select-popover"] .MuiListItem-root').eq(1).click();